+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.5.0dev (master) ##
 
+- New Rules: ```title-match-regex```
+
 ## v0.4.1 (2015-09-19) ##
 
 - Internal fix: added missing comma to setup.py which prevented pypi upload